move `DB` folder to `src`
